requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
nltk==3.8.1
streamlit==1.32.0
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
huggingface-hub==0.14.1
ctransformers==0.2.27
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2